+++ CHANGELOG.md
@@ -39,6 +39,7 @@ All notable changes to this project will be documented in this file.
 - *(pdf)* Added redis adapter to cache the topic book extraction
 - *(pdf)* Added route to get specific inputs (pages) based on a temporary memory 30min
 - *(pdf)* Added dataclass to improve the pdf reading
+- *(pdf)* Organized the env lib for the Adapter for PDF
 
 ### ⚙️ Miscellaneous Tasks